EP11

+++ server/controllers/product.js
@@ -1,5 +1,16 @@
 const prisma = require("../config/prisma")
 const cloudinary = require('cloudinary').v2;
+
+
+// Configuration
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
+});
+
+
+
 exports.create = async (req, res) => {
     try {
         // code
@@ -46,7 +57,6 @@ exports.list = async (req, res) => {
         res.status(500).json({ message: "Server error" })
     }
 }
-
 exports.read = async (req, res) => {
     try {
         // code
@@ -108,9 +118,29 @@ exports.remove = async (req, res) => {
     try {
         // code
         const { id } = req.params
-
         // หนังชีวิต 
-
+        // Step 1 ค้นหาสินค้า include images
+        const product = await prisma.product.findFirst({
+            where: { id: Number(id) },
+            include: { images: true }
+        })
+        if (!product) {
+            return res.status(400).json({ message: 'Product not found!!' })
+        }
+        // console.log(product)
+        // Step 2 Promise ลบรูปภาพใน cloud ลบแบบ รอฉันด้วย
+        const deletedImage = product.images
+        .map((image)=>
+        new Promise((resolve,reject)=>{
+            // ลบจาก cloud
+            cloudinary.uploader.destroy(image.public_id,(error,result)=>{
+                if(error) reject(error)
+                else resolve(result)
+            })
+        })
+        )
+        await Promise.all(deletedImage)
+        // Step 3 ลบสินค้า
         await prisma.product.delete({
             where: {
                 id: Number(id)
@@ -231,12 +261,7 @@ exports.searchFilters = async (req, res) => {
 }
 
 
-// Configuration
-cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET,
-});
+
